improved callbacks docs
